Init adding known mutations
