+++ gff2neo/data/drugbank/README.md
@@ -1,0 +1,3 @@
+## DrugBank Data
+* `all_target_polypeptide_ids.csv` is from [DrugBank](https://www.drugbank.ca/releases/latest#protein-identifiers). Released: 2018-04-02.
+* `gff2neo/data/drugbank/drugbank_vocabulary.csv` from [doi:10.1038/ng.3767](https://www.nature.com/articles/ng.3767#integrated-supplementary-information).
