+++ gff2neo/data/mutations/README.md
@@ -1,0 +1,3 @@
+## Known mutations
+* `drdb.txt` is from [TBProfiler](https://github.com/jodyphelan/TBProfiler/tree/master/db).
+* `ng.3767-s3.xlsx` from [doi:10.1038/ng.3767](https://www.nature.com/articles/ng.3767#integrated-supplementary-information).
