+++ gff2neo/quickgo.py
@@ -30,9 +30,14 @@ BASE_URL = "https://www.ebi.ac.uk/QuickGO/services"
 
 
 def query_quickgo(go_term_ids):
+    """
+    Retrieve information given a GO identifiers.
+    :param str go_term_ids:
+    :return:
+    """
     s = Session()
     s.mount(BASE_URL, HTTPAdapter(
-        max_retries=Retry(total=5, status_forcelist=[500])
+        max_retries=Retry(total=5, backoff_factor=0.2, status_forcelist=[500])
     ))
     url = BASE_URL + "/ontology/go/terms/"
     headers = {'Accept': 'application/json'}
