+++ requirements.txt
@@ -49,3 +49,4 @@ tqdm==4.15.0
 urllib3==1.22
 wrapt==1.10.10
 xmltodict==0.11.0
+xlrd
