+++ test/test_quickgo.py
@@ -1,3 +1,8 @@
+"""
+Testing quickgo module
+"""
+import pytest
+
 from gff2neo.quickgo import query_quickgo
 
 
@@ -11,7 +16,13 @@ from gff2neo.quickgo import query_quickgo
 #     result = fetch_quick_go_data(quick_go, "GO:0005886")
 #     assert isinstance(result, list) is True
 
+@pytest.mark.parametrize("test_input,expected", [
+    (query_quickgo("GO:0005576,GO:0005618,GO:0016021").status_code, 200),
+    (query_quickgo("GO:0003993,GO:0005886,GO:0046872").status_code, 200),
+    (query_quickgo("GO:0003824,GO:0005886,GO:0016021,GO:0046872,GO:0051539").status_code, 200),
+    (query_quickgo(
+        "GO:0004160,GO:0005886,GO:0009082,GO:0009097,GO:0009099,GO:0040007,GO:0046872,GO:0051539").status_code, 200),
 
-def test_query_quickgo():
-    result = query_quickgo("GO:0004160,GO:0005886,GO:0009082,GO:0009097,GO:0009099,GO:0040007,GO:0046872,GO:0051539")
-    assert result.status_code == 200
+])
+def test_query_quickgo(test_input, expected):
+    assert isinstance(test_input, expected) is True
